requirements.txt
azure-cognitiveservices-speech
regex
requests
sounddevice
soundfile
tiktoken
openai>=1.40.0
nvidia-riva-client>=2.14.0
newspaper3k
llama-hub
llama-index-llms-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-embeddings-openai
llama-index-agent-openai
llama-index-tools-bing-search
llama-index-tools-weather
llama-index-tools-google
llama-index-tools-wikipedia
pyowm
llama-index
langchain>=0.0.329
# mediapipe
pvporcupine
# pyaudio
# struct
webrtcvad
cohere
groq
google.generativeai
supabase
whisper
wget
youtube_transcript_api
moviepy
ffmpeg-python
pytube
pinecone-client
lxml_html_clean
gpt_researcher
numpy
litellm
pyyaml
